Clean and rename ripple LLC family into xrp family
