+++ libs/ledger-live-common/.unimportedrc.json
@@ -120,7 +120,6 @@
     "src/families/polkadot/logic.ts",
     "src/families/polkadot/react.ts",
     "src/families/polkadot/types.ts",
-    "src/families/ripple/types.ts",
     "src/families/solana/banner.ts",
     "src/families/solana/logic.ts",
     "src/families/solana/react.ts",
@@ -134,6 +133,10 @@
     "src/families/tezos/types.ts",
     "src/families/tron/react.ts",
     "src/families/tron/types.ts",
+    "src/families/xrp/bridge.integration.test.ts",
+    "src/families/xrp/platformAdapter.test.ts",
+    "src/families/xrp/types.ts",
+    "src/families/xrp/walletApiAdapter.test.ts",
     "src/featureFlags/FeatureFlagsContext.ts",
     "src/featureFlags/groupedFeatures.ts",
     "src/featureFlags/index.ts",
@@ -280,8 +283,15 @@
     "src/wallet-api/types.ts",
     "src/wallet-api/version.ts"
   ],
-  "extensions": [".ts", ".js", ".jsx", ".tsx"],
-  "ignorePatterns": ["**/node_modules/**"],
+  "extensions": [
+    ".ts",
+    ".js",
+    ".jsx",
+    ".tsx"
+  ],
+  "ignorePatterns": [
+    "**/node_modules/**"
+  ],
   "ignoreUnimported": [
     "lib-es/__tests__",
     "lib-es/__tests__/accounts",
@@ -945,18 +955,18 @@
     "lib-es/families/polkadot/walletApiAdapter.test.d.ts.map",
     "lib-es/families/polkadot/walletApiAdapter.test.js",
     "lib-es/families/polkadot/walletApiAdapter.test.js.map",
-    "lib-es/families/ripple/bridge.integration.test.d.ts",
-    "lib-es/families/ripple/bridge.integration.test.d.ts.map",
-    "lib-es/families/ripple/bridge.integration.test.js",
-    "lib-es/families/ripple/bridge.integration.test.js.map",
-    "lib-es/families/ripple/platformAdapter.test.d.ts",
-    "lib-es/families/ripple/platformAdapter.test.d.ts.map",
-    "lib-es/families/ripple/platformAdapter.test.js",
-    "lib-es/families/ripple/platformAdapter.test.js.map",
-    "lib-es/families/ripple/walletApiAdapter.test.d.ts",
-    "lib-es/families/ripple/walletApiAdapter.test.d.ts.map",
-    "lib-es/families/ripple/walletApiAdapter.test.js",
-    "lib-es/families/ripple/walletApiAdapter.test.js.map",
+    "lib-es/families/xrp/bridge.integration.test.d.ts",
+    "lib-es/families/xrp/bridge.integration.test.d.ts.map",
+    "lib-es/families/xrp/bridge.integration.test.js",
+    "lib-es/families/xrp/bridge.integration.test.js.map",
+    "lib-es/families/xrp/platformAdapter.test.d.ts",
+    "lib-es/families/xrp/platformAdapter.test.d.ts.map",
+    "lib-es/families/xrp/platformAdapter.test.js",
+    "lib-es/families/xrp/platformAdapter.test.js.map",
+    "lib-es/families/xrp/walletApiAdapter.test.d.ts",
+    "lib-es/families/xrp/walletApiAdapter.test.d.ts.map",
+    "lib-es/families/xrp/walletApiAdapter.test.js",
+    "lib-es/families/xrp/walletApiAdapter.test.js.map",
     "lib-es/families/solana/banner.test.d.ts",
     "lib-es/families/solana/banner.test.d.ts.map",
     "lib-es/families/solana/banner.test.js",
@@ -1893,18 +1903,18 @@
     "lib/families/polkadot/walletApiAdapter.test.d.ts.map",
     "lib/families/polkadot/walletApiAdapter.test.js",
     "lib/families/polkadot/walletApiAdapter.test.js.map",
-    "lib/families/ripple/bridge.integration.test.d.ts",
-    "lib/families/ripple/bridge.integration.test.d.ts.map",
-    "lib/families/ripple/bridge.integration.test.js",
-    "lib/families/ripple/bridge.integration.test.js.map",
-    "lib/families/ripple/platformAdapter.test.d.ts",
-    "lib/families/ripple/platformAdapter.test.d.ts.map",
-    "lib/families/ripple/platformAdapter.test.js",
-    "lib/families/ripple/platformAdapter.test.js.map",
-    "lib/families/ripple/walletApiAdapter.test.d.ts",
-    "lib/families/ripple/walletApiAdapter.test.d.ts.map",
-    "lib/families/ripple/walletApiAdapter.test.js",
-    "lib/families/ripple/walletApiAdapter.test.js.map",
+    "lib/families/xrp/bridge.integration.test.d.ts",
+    "lib/families/xrp/bridge.integration.test.d.ts.map",
+    "lib/families/xrp/bridge.integration.test.js",
+    "lib/families/xrp/bridge.integration.test.js.map",
+    "lib/families/xrp/platformAdapter.test.d.ts",
+    "lib/families/xrp/platformAdapter.test.d.ts.map",
+    "lib/families/xrp/platformAdapter.test.js",
+    "lib/families/xrp/platformAdapter.test.js.map",
+    "lib/families/xrp/walletApiAdapter.test.d.ts",
+    "lib/families/xrp/walletApiAdapter.test.d.ts.map",
+    "lib/families/xrp/walletApiAdapter.test.js",
+    "lib/families/xrp/walletApiAdapter.test.js.map",
     "lib/families/solana/banner.test.d.ts",
     "lib/families/solana/banner.test.d.ts.map",
     "lib/families/solana/banner.test.js",
@@ -2619,9 +2629,6 @@
     "src/families/polkadot/bridge.integration.test.ts",
     "src/families/polkadot/platformAdapter.test.ts",
     "src/families/polkadot/walletApiAdapter.test.ts",
-    "src/families/ripple/bridge.integration.test.ts",
-    "src/families/ripple/platformAdapter.test.ts",
-    "src/families/ripple/walletApiAdapter.test.ts",
     "src/families/solana/__snapshots__/bridge.integration.test.ts.snap",
     "src/families/solana/bridge/mock.ts",
     "src/families/solana/bridge/mock-data.ts",
