+++ libs/ledger-live-common/src/config/sharedConfig.ts
@@ -13,7 +13,7 @@ import { hederaConfig } from "../families/hedera/config";
 import { internetComputerConfig } from "../families/internet_computer/config";
 import { nearConfig } from "../families/near/config";
 import { polkadotConfig } from "../families/polkadot/config";
-import { rippleConfig } from "../families/ripple/config";
+import { xrpConfig } from "../families/xrp/config";
 import { solanaConfig } from "../families/solana/config";
 import { stacksConfig } from "../families/stacks/config";
 import { stellarConfig } from "../families/stellar/config";
@@ -54,7 +54,7 @@ export const liveConfig: ConfigSchema = {
   ...internetComputerConfig,
   ...nearConfig,
   ...polkadotConfig,
-  ...rippleConfig,
+  ...xrpConfig,
   ...solanaConfig,
   ...stacksConfig,
   ...stellarConfig,
