+++ libs/ledger-live-common/src/families/xrp/types.ts
@@ -1,0 +1,2 @@
+// Encapsulate for LLD et LLM
+export * from "@ledgerhq/coin-xrp/types";
