+++ libs/ledger-live-common/src/families/xrp/walletApiAdapter.ts
@@ -1,7 +1,6 @@
 import { RippleTransaction as WalletAPIRippleTransaction } from "@ledgerhq/wallet-api-core";
 import { AreFeesProvided, GetWalletAPITransactionSignFlowInfos } from "../../wallet-api/types";
-
-import { Transaction } from "./types";
+import { Transaction } from "@ledgerhq/coin-xrp/types";
 
 const CAN_EDIT_FEES = true;
